+++ CHANGELOG.md
@@ -504,6 +504,7 @@
   - /clock_unknown
 * Day cycle clocks can now have their hour length and current hour be modified via /clock_set
 * Day cycle clock unpauses now take place as soon as processed rather than at most 1 second after being processed
+* Added /zone_mode to set up the gamemode of a zone. Players part of an area in that zone with compatible clients will automatically change to their gamemode's version of their theme. 
 * Players in a party that are sneaking may now see each other via /getarea and similar. Players in the party not sneaking, or players sneaking not part of the party may not see these players
 * Changed wording of GM login notifications, /minimap, attempting to access a locked passage, and talking in an area whose IC chat is locked
 * Notifications are now sent if a mod via /switch forces a target off their character (e.g. mod using /switch) to the mod, the target, and other officers in the server